MoveIn/shared/requirements.txt
pydantic>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
fastapi>=0.104.0
